refactor: rename to SearchSelect
